change Tcl_Size to long for <8.7

+++ xpa/tcl.c
@@ -12,7 +12,7 @@
    which was introduced in Tcl 8.7 and 9.
 */
 #ifndef Tcl_Size
-typedef int Tcl_Size;
+typedef long Tcl_Size;
 #endif
 
 /*
